Credorax: Support referral CFT transactions
Added support for the referral CFT transaction type on the Credorax
gateway (transaction type [34]).

To simplify this new transaction type, it is selected by passing a new
`referral_cft` gateway specific field while requesting a `refund`
transaction. Although the referral CFT transaction defaults to a credit
request when it has not been enabled on the Credorax gateway
integration, the referral CFT requires passing a `g2` value, i.e.,
"Previous Response ID". Therefore, because this transaction type
requires a reference authorization from a previous transaction, its
implmentation was more similar to a refund than a credit.

Because Spreedly's current Credorax gateway integration does not have
the ability to make referral CFT transactions, the remote and unit tests
for this new transaction type are currently failing. These have been
left as failing to indicate that they have not been confirmed to work.
Although an attempt could be made to guess the gateway's response for
the unit tests, the tests have been left with 'RESPONSE NEEDED' to
better and more clearly indicate the unknowns.

CE-203

Unit:
60 tests, 285 assertions, 2 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
96.6667% passed

Remote:
37 tests, 72 assertions, 19 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
48.6486% passed

About half of the remote tests are currently failing with the message
"Do not Honour". This usually happens in the afternoons but has
recently been happening even in the mornings. How to fix these errors
is currently unknown.